transports: merge common error handling.
